Class extract
