+++ src/SyncGitHubReleases/Program.cs
@@ -67,7 +67,7 @@ namespace SyncGitHubReleases
                 if (token == null)
                 {
                     Console.WriteLine("No GitHub token specified so assuming URL points to existing RELEASES file");
-                    await SyncFromStandardHttp(new Uri(repoUrl), releasesDirectoryInfo);
+                    await new StandardHttpSyncer(new Uri(repoUrl)).Sync(releasesDirectoryInfo);
                 }
                 else
                 {
@@ -144,60 +144,7 @@ namespace SyncGitHubReleases
 
             ReleaseEntry.WriteReleaseFile(entries, Path.Combine(releaseDirectoryInfo.FullName, "RELEASES"));
         }
-
-        private async Task SyncFromStandardHttp(Uri repoUrl, DirectoryInfo releasesDir)
-        {
-            var releasesIndex = await DownloadReleasesIndex(repoUrl);
-
-            File.WriteAllText(Path.Combine(releasesDir.FullName, "RELEASES"), releasesIndex);
-
-            var releasesToDownload = ReleaseEntry.ParseReleaseFile(releasesIndex)
-                .Where(x => !x.IsDelta)
-                .OrderByDescending(x => x.Version)
-                .Take(2)
-                .Select(x => new
-                             {
-                                 LocalPath = Path.Combine(releasesDir.FullName, x.Filename),
-                                 RemoteUrl = new Uri(repoUrl, x.Filename)
-                             }
-                );
-
-            foreach (var releaseToDownload in releasesToDownload)
-                await DownloadRelease(releaseToDownload.LocalPath, releaseToDownload.RemoteUrl);
-        }
-
-        private async Task<string> DownloadReleasesIndex(Uri repoUrl)
-        {
-            var uri = new Uri(repoUrl, "RELEASES");
-
-            Console.WriteLine("Trying to download RELEASES index from {0}", uri);
-
-            using (HttpClient client = new HttpClient())
-            {
-                return await client.GetStringAsync(uri);
-            }
-        }
-
-        private async Task DownloadRelease(string localPath, Uri remoteUrl)
-        {
-            if (File.Exists(localPath))
-            {
-                Console.WriteLine("Skipping this release as existing file found at: {0}", localPath);
-                return;
-            }
-
-            using (HttpClient client = new HttpClient())
-            {
-                using (var localStream = File.Create(localPath))
-                {
-                    using (var remoteStream = await client.GetStreamAsync(remoteUrl))
-                    {
-                        await remoteStream.CopyToAsync(localStream);
-                    }
-                }
-            }
-        }
-
+        
         public void ShowHelp()
         {
             opts.WriteOptionDescriptions(Console.Out);
