+++ src/SyncGitHubReleases/SyncGitHubReleases.csproj
@@ -66,6 +66,7 @@
     <Compile Include="Mono.Options\Options.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StandardHttpSyncer.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
